Revert "Plan 9 from Bell Labs Fourth Edition 2005-12-26"
This reverts commit 74f16c8187aab477889167f2422d0597b1b7d0ff.

+++ sys/src/cmd/fossil/dat.h
@@ -214,7 +214,7 @@ struct Block {
 	Cache *c;
 	int ref;
 	int nlock;
-	uintptr	pc;		/* pc that fetched this block from the cache */
+	ulong	pc;		/* pc that fetched this block from the cache */
 
 	VtLock *lk;
